move slackHelper
